requirements.txt
# Core Libraries
Flask
gunicorn
pydantic
python-dotenv
requests

# Google AI Integration
langchain
langchain-core
langchain-google-genai

# Image Processing
Pillow
